+++ src/commands/init.ts
@@ -10,9 +10,10 @@ import untildify from 'untildify';
 import Command from '../base-command';
 import DockerComposeTemplate from '../common/docker-compose/template';
 import { AccountUtils } from '../common/utils/account';
+import { InterfaceSpecV1 } from '../dependency-manager/src/spec/common/interface-v1';
 import { ComponentConfigV1 } from '../dependency-manager/src/spec/component/v1';
 import { BuildSpecV1, ServiceVolumeV1 } from '../dependency-manager/src/spec/resource/v1';
-import { InterfaceSpecV1, ServiceConfigV1 } from '../dependency-manager/src/spec/service/v1';
+import { ServiceConfigV1 } from '../dependency-manager/src/spec/service/v1';
 
 export abstract class InitCommand extends Command {
   auth_required() {
