+++ src/dependency-manager/src/graph/node/index.ts
@@ -1,5 +1,5 @@
 import { Type } from 'class-transformer';
-import { InterfaceSpec } from '../../spec/service/base';
+import { InterfaceSpec } from '../../spec/common/interface-spec';
 import { Dictionary } from '../../utils/dictionary';
 import { DependencyState } from '../state';
