+++ src/dependency-manager/src/index.ts
@@ -5,6 +5,9 @@ export * from './graph/node';
 export * from './graph/node/service';
 export * from './graph/node/task';
 export * from './spec/base-spec';
+export * from './spec/common/interface-spec';
+export * from './spec/common/liveness-probe-spec';
+export * from './spec/common/parameter-spec';
 export * from './spec/component/base';
 export * from './spec/component/builder';
 export * from './spec/environment/base';
