+++ src/dependency-manager/src/manager.ts
@@ -7,11 +7,11 @@ import ServiceEdge from './graph/edge/service';
 import GatewayNode from './graph/node/gateway';
 import InterfacesNode from './graph/node/interfaces';
 import { TaskNode } from './graph/node/task';
+import { InterfaceSpec } from './spec/common/interface-spec';
 import { ComponentConfig } from './spec/component/base';
 import { ComponentConfigBuilder } from './spec/component/builder';
 import { EnvironmentConfig } from './spec/environment/base';
 import { EnvironmentConfigBuilder } from './spec/environment/builder';
-import { InterfaceSpec } from './spec/service/base';
 import { Dictionary } from './utils/dictionary';
 import { flattenValidationErrors, ValidationErrors } from './utils/errors';
 import { escapeJSON, interpolateString, normalizeInterpolation, prefixExpressions, removePrefixForExpressions, replaceBrackets } from './utils/interpolation';
