+++ src/dependency-manager/src/spec/common/interface-spec.ts
@@ -1,0 +1,8 @@
+export interface InterfaceSpec {
+  description?: string;
+  host?: string;
+  port: string;
+  url?: string;
+  protocol?: string;
+  domains?: string[];
+}
