+++ src/dependency-manager/src/spec/common/parameter-spec.ts
@@ -1,0 +1,10 @@
+
+export type ParameterValue = string | number | boolean | undefined;
+
+export interface ParameterDefinitionSpec {
+  required?: boolean;
+  description?: string;
+  default?: ParameterValue;
+}
+
+export type ParameterValueSpec = ParameterValue | ParameterDefinitionSpec;
