+++ src/dependency-manager/src/spec/service/base.ts
@@ -1,25 +1,7 @@
 import { Dictionary } from '../../utils/dictionary';
+import { InterfaceSpec } from '../common/interface-spec';
+import { ServiceLivenessProbe } from '../common/liveness-probe-spec';
 import { ResourceConfig } from '../resource/base';
-
-export interface InterfaceSpec {
-  description?: string;
-  host?: string;
-  port: string;
-  url?: string;
-  protocol?: string;
-  domains?: string[];
-}
-
-export interface ServiceLivenessProbe {
-  success_threshold?: string;
-  failure_threshold?: string;
-  timeout?: string;
-  path?: string;
-  interval?: string;
-  command?: string[];
-  port?: string;
-  initial_delay?: string;
-}
 
 export interface ServiceConfig extends ResourceConfig {
   getInterfaces(): Dictionary<InterfaceSpec>;
